requirements.txt
PyYAML==5.4.1
SQLAlchemy==1.4.22
PyMySQL==1.0.2
typer==0.3.2

mysql==0.0.3
mysql-connector==2.2.9
alive_progress==1.6.2
mysqlclient==2.0.3
